requirements.txt
asgiref==3.7.2
astroid==3.2.2
backports.zoneinfo==0.2.1
dill==0.3.8
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.8.0
pylint==3.2.3
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.9.0
uritemplate==4.1.1
